requirements.txt
scrapy>=2.11.0
pymongo>=4.6.1
requests>=2.31.0
parsel>=1.8.1
cryptography>=42.0.2
certifi>=2024.2.2
urllib3>=2.2.0
pydantic>=2.6.1
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
apache-airflow>=2.8.1
kafka-python>=2.0.2
pytest>=8.0.0
flake8>=7.0.0
mypy>=1.8.0
black>=24.1.1
isort>=5.13.2
